requirements.txt
streamlit
openai
pandas
matplotlib
seaborn
python-dotenv
requests
PyPDF2